+++ src/fobi/contrib/plugins/form_elements/fields/select_multiple_model_objects/__init__.py
@@ -1,6 +1,6 @@
 __title__ = 'fobi.contrib.plugins.form_elements.fields.select_multiple_model_objects'
 __author__ = 'Artur Barseghyan <artur.barseghyan@gmail.com>'
-__copyright__ = 'Copyright (c) 2014 Artur Barseghyan'
+__copyright__ = 'Copyright (c) 2014-2015 Artur Barseghyan'
 __license__ = 'GPL 2.0/LGPL 2.1'
 __all__ = ('default_app_config', 'UID',)