+++ src/fobi/contrib/plugins/form_elements/fields/select_multiple_model_objects/conf.py
@@ -1,6 +1,6 @@
 __title__ = 'fobi.contrib.plugins.form_elements.fields.select_multiple_model_objects.conf'
 __author__ = 'Artur Barseghyan <artur.barseghyan@gmail.com>'
-__copyright__ = 'Copyright (c) 2014 Artur Barseghyan'
+__copyright__ = 'Copyright (c) 2014-2015 Artur Barseghyan'
 __license__ = 'GPL 2.0/LGPL 2.1'
 __all__ = ('get_setting',)
 
@@ -10,13 +10,15 @@ from fobi.contrib.plugins.form_elements.fields.select_model_object import defaul
 
 def get_setting(setting, override=None):
     """
-    Get a setting from `fobi.contrib.plugins.form_elements.fields.select_multiple_model_objects`
+    Get a setting from
+    `fobi.contrib.plugins.form_elements.fields.select_multiple_model_objects`
     conf module, falling back to the default.
 
     If override is not None, it will be used instead of the setting.
 
     :param setting: String with setting name
-    :param override: Value to use when no setting is available. Defaults to None.
+    :param override: Value to use when no setting is available. Defaults
+        to None.
     :return: Setting value.
     """
     if override is not None: