+++ src/fobi/views.py
@@ -884,6 +884,9 @@ def view_form_entry(request, form_entry_slug, theme=None, template_name=None):
                                 form=form, stage=CALLBACK_FORM_INVALID)
 
     else:
+        # Providing initial form data by feeding entire GET dictionary
+        # to the form, if ``GET_PARAM_INITIAL_DATA`` is present in the
+        # GET.
         kwargs = {}
         if GET_PARAM_INITIAL_DATA in request.GET:
             kwargs = {'initial': request.GET}